refactored files ready for expansion
